:sparkles: Finished up server creation and deletion
